// src/api/generated/schemas/ticketUpdateRequestStatus.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */

export type TicketUpdateRequestStatus = typeof TicketUpdateRequestStatus[keyof typeof TicketUpdateRequestStatus];


export const TicketUpdateRequestStatus = {
  PENDING: 'PENDING',
  AWAITING_CUSTOMER: 'AWAITING_CUSTOMER',
  AWAITING_DEVELOPMENT: 'AWAITING_DEVELOPMENT',
  COMPLETED: 'COMPLETED',
  CANCELLED: 'CANCELLED',
} as const;