// src/api/generated/schemas/ticketResponse.ts
/**
 * Generated by orval v8.8.1 🍺
 * Do not edit manually.
 * WorkLog API
 * API para gerenciamento de tickets e interações de suporte ao cliente. Autenticação via cookie HttpOnly `worklog_access` emitido pelo endpoint `POST /worklog/auth/login` (o browser anexa o cookie automaticamente em chamadas subsequentes).
 * OpenAPI spec version: v1.0.0
 */
import type { ClientSummary } from './clientSummary';
import type { SystemResponse } from './systemResponse';
import type { TicketResponsePriority } from './ticketResponsePriority';
import type { TicketResponseStatus } from './ticketResponseStatus';
import type { UserSummary } from './userSummary';

export interface TicketResponse {
  publicId?: string;
  title?: string;
  description?: string;
  solution?: string;
  status?: TicketResponseStatus;
  priority?: TicketResponsePriority;
  createdAt?: string;
  completedAt?: string;
  updatedAt?: string;
  client?: ClientSummary;
  system?: SystemResponse;
  user?: UserSummary;
}